+++ packages/shared/routes/dashboard/wallet/Wallet.svelte
@@ -45,7 +45,7 @@
         updateBalanceOverview,
         wallet,
     } from 'shared/lib/wallet'
-    import { initialiseListeners } from 'shared/lib/listeners'
+    import { initialiseListeners } from 'lib/walletApiListeners'
     import { onMount, setContext } from 'svelte'
     import { derived, Readable } from 'svelte/store'
     import { AccountActions, AccountBalance, AccountHistory, AccountNavigation, BarChart, LineChart } from './views/'
